backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
redis==5.0.1
aiofiles==23.2.1
python-multipart==0.0.6
pillow==10.1.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-multipart==0.0.6

# Cloud Storage
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0
boto3==1.34.9
botocore==1.34.9

# Database for user data
alembic==1.13.1
asyncpg==0.29.0

# Utilities
watchdog==3.0.0
httpx==0.25.2
unidecode==1.3.7
requests==2.31.0
tqdm==4.66.1
urllib3<2.1,>=1.25.4
importlib-metadata>=4.0.0; python_version < "3.10"  # Backport for Python 3.9
qrcode[pil]==7.4.2

# Email (optional - only needed for SMTP fallback, not required for AWS SES)
# aiosmtplib==3.0.1
